Separated Services based on functions to 4 different service: User, Unit, Recipe, Diet
