+++ backend/src/main/java/com/codecool/chilibeans/controller/UnitController.java
@@ -2,7 +2,7 @@ package com.codecool.chilibeans.controller;
 
 import com.codecool.chilibeans.controller.dto.unit.NewUnitDTO;
 import com.codecool.chilibeans.controller.dto.unit.UnitDTO;
-import com.codecool.chilibeans.service.ChiliServiceInterface;
+import com.codecool.chilibeans.service.UnitService;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Set;
@@ -12,35 +12,35 @@ import java.util.UUID;
 @RequestMapping("/api/units")
 public class UnitController {
 
-    ChiliServiceInterface chiliService;
+    UnitService unitService;
 
-    public UnitController(ChiliServiceInterface chiliService) {
-        this.chiliService = chiliService;
+    public UnitController(UnitService unitService) {
+        this.unitService = unitService;
     }
 
     @GetMapping("")
     public Set<UnitDTO> getAllUnits() {
-        return chiliService.getAllUnits();
+        return unitService.getAllUnits();
     }
 
     @GetMapping("/{id}")
     public UnitDTO getUnitById(@PathVariable("id") UUID uuid) {
-        return chiliService.getUnitById(uuid);
+        return unitService.getUnitById(uuid);
     }
 
     @PostMapping("")
     public UnitDTO createUnit(@RequestBody NewUnitDTO newUnitDTO) {
-        return chiliService.createNewUnit(newUnitDTO);
+        return unitService.createNewUnit(newUnitDTO);
     }
 
     @DeleteMapping("/{id}")
     public UnitDTO deleteUnitById(@PathVariable("id") UUID uuid) {
-        return chiliService.deleteUnitById(uuid);
+        return unitService.deleteUnitById(uuid);
     }
 
     @PatchMapping("/{id}")
     public UnitDTO updateUnitById(@PathVariable("id") UUID uuid, @RequestBody UnitDTO unitDTO) {
-        return chiliService.updateUnitById(uuid, unitDTO);
+        return unitService.updateUnitById(uuid, unitDTO);
     }
 
 }
