+++ backend/src/main/java/com/codecool/chilibeans/service/RecipeService.java
@@ -1,0 +1,21 @@
+package com.codecool.chilibeans.service;
+
+import com.codecool.chilibeans.controller.dto.recipe.NewRecipeDTO;
+import com.codecool.chilibeans.controller.dto.recipe.RecipeDTO;
+
+import java.util.Set;
+import java.util.UUID;
+
+public interface RecipeService {
+
+    Set<RecipeDTO> getAllRecipes();
+
+    RecipeDTO getRecipeById(UUID id);
+
+    RecipeDTO createNewRecipe(NewRecipeDTO newRecipeDTO);
+
+    RecipeDTO updateRecipe(UUID id, RecipeDTO recipeDTO);
+
+    boolean deleteRecipeById(UUID id);
+
+}
