+++ backend/src/main/java/com/codecool/chilibeans/service/UnitService.java
@@ -1,0 +1,23 @@
+package com.codecool.chilibeans.service;
+
+import com.codecool.chilibeans.controller.dto.unit.NewUnitDTO;
+import com.codecool.chilibeans.controller.dto.unit.UnitDTO;
+import org.springframework.stereotype.Service;
+
+import java.util.Set;
+import java.util.UUID;
+
+@Service
+public interface UnitService {
+
+    Set<UnitDTO> getAllUnits();
+
+    UnitDTO createNewUnit(NewUnitDTO newUnitDTO);
+
+    UnitDTO getUnitById(UUID uuid);
+
+    UnitDTO deleteUnitById(UUID uuid);
+
+    UnitDTO updateUnitById(UUID uuid, UnitDTO unitDTO);
+
+}
